Fixed bug with toolchain install/uninstall

+++ simpleiot/cli/buildtool/toolchain.py
@@ -69,30 +69,41 @@ class Toolchain:
         key = self._make_key(manufacturer, processor, opsys, version)
         print("Deletes configuration data for this platform--after install")
 
-    def install_alias(self, base, alias):
+    def install_alias(self, base, alias, verbose):
         params = self._resolve_alias(alias)
         self.install(base,
                      params["man"],
                      params["pro"],
                      params["ops"],
                      params["ver"],
-                     params["loc"])
+                     params["loc"], verbose)
 
 
-    def install(self, base, manufacturer, processor, opsys, version, location):
+    def install(self, base, manufacturer, processor, opsys, version, location, verbose=False):
         try:
+            if verbose:
+                print(f"+ Looking up build tool for: mfr: {manufacturer} proc: {processor} os: {opsys} ver: {version} verbose: {verbose}")
             key = self._make_key(manufacturer, processor, opsys, version)
+            if verbose:
+                print(f"+ Generated install hash key: {key}")
             installer = self.toolchain_list.get(key, None)
+            if verbose:
+                print(f"+ Installer instance is: {str(installer)}")
+
             if installer:
                 install_dir = installer.install_path(base)
+                if verbose:
+                    print(f"+ Install directory is: {install_dir}")
 
                 opsys = platform.system()
+                if verbose:
+                    print(f"+ Runtime OS is: {opsys}")
                 if opsys == "Darwin":
-                    installer.install_mac(install_dir, location)
+                    installer.install_mac(install_dir, location, verbose)
                 elif opsys == "Windows":
-                    installer.install_windows(install_dir, location)
+                    installer.install_windows(install_dir, location, verbose)
                 elif opsys == "Linux":
-                    installer.install_unix(install_dir, location)
+                    installer.install_unix(install_dir, location, verbose)
                 else:
                     print(f"ERROR: operating system not supported")
             else:
@@ -104,26 +115,39 @@ class Toolchain:
             exit(1)
 
 
-    def uninstall(self, base, manufacturer, processor, opsys, version, location):
+    def uninstall(self, base, manufacturer, processor, opsys, version, location, verbose=False):
         try:
+            if verbose:
+                print(f"+ Looking up build tool for: mfr: {manufacturer} proc: {processor} os: {opsys} ver: {version}")
             key = self._make_key(manufacturer, processor, opsys, version)
+            if verbose:
+                print(f"+ Generated install hash key: {key}")
             installer = self.toolchain_list.get(key, None)
             if installer:
                 install_dir = installer.install_path(base)
+                if verbose:
+                    print(f"+ Install directory is: {install_dir}")
 
                 opsys = platform.system()
+                if verbose:
+                    print(f"+ Runtime OS is: {opsys}")
                 if opsys == "Darwin":
-                    installer.uninstall_mac(install_dir, location)
+                    installer.uninstall_mac(install_dir, location, verbose)
                 elif opsys == "Windows":
-                    installer.uninstall_windows(install_dir, location)
+                    installer.uninstall_windows(install_dir, location, verbose)
                 elif opsys == "Linux":
-                    installer.uninstall_unix(install_dir, location)
+                    installer.uninstall_unix(install_dir, location, verbose)
                 else:
                     print(f"ERROR: operating system not supported")
 
-                # Also remove the directory in .simpleiot
+                # Also remove the directory in .simpleiot if it still exists
                 #
-                shutil.rmtree(install_dir)
+                if os.path.exists(install_dir):
+                    if verbose:
+                        print(f"+ Cleaning remnant directory {install_dir}")
+                    shutil.rmtree(install_dir)
+
+                print("Done.")
             else:
                 print(f"ERROR: unable to uninstall toolchain for specified device")
                 exit(1)
@@ -131,23 +155,32 @@ class Toolchain:
             print(f"ERROR uninstalling tool: {str(e)}")
             exit(1)
 
-    def reset(self, base, manufacturer, processor, opsys, version, location):
+    def reset(self, base, manufacturer, processor, opsys, version, location, verbose=False):
         try:
             print("Loading defaults for toolchain...")
+            if verbose:
+                print(f"+ Looking up build tool for: mfr: {manufacturer} proc: {processor} os: {opsys} ver: {version}")
             key = self._make_key(manufacturer, processor, opsys, version)
+            if verbose:
+                print(f"+ Generated install hash key: {key}")
+
             installer = self.toolchain_list.get(key, None)
             if installer:
                 opsys = platform.system()
+                if verbose:
+                    print(f"+ Runtime OS is: {opsys}")
                 if opsys == "Darwin":
-                    installer.reset_mac(base, location)
+                    installer.reset_mac(base, location, verbose)
                 elif opsys == "Windows":
                     # For windows, we install the tool in a custom directory under ~/.simpleiot/_toolchain/{hash}
                     install_path = Path(base) / key
-                    installer.reset_windows(install_path, location)
+                    installer.reset_windows(install_path, location, verbose)
                 elif opsys == "Linux":
-                    installer.reset_unix(base, location)
+                    installer.reset_unix(base, location, verbose)
                 else:
                     print(f"ERROR: operating system not supported")
+
+                print("Done.")
             else:
                 print(f"ERROR: No toolchain found. Unable to reset toolchain for specified device")
                 exit(1)
@@ -155,12 +188,21 @@ class Toolchain:
             print(f"ERROR resetting tool to default settings: {str(e)}")
             exit(1)
 
-    def build(self, base, manufacturer, processor, opsys, version, location, dirpath, command):
+    def build(self, base, manufacturer, processor, opsys, version, location, dirpath, command, verbose=False):
         try:
+            if verbose:
+                print(f"+ Looking up build tool for: mfr: {manufacturer} proc: {processor} os: {opsys} ver: {version}")
             key = self._make_key(manufacturer, processor, opsys, version)
+            if verbose:
+                print(f"+ Generated install hash key: {key}")
+
             installer = self.toolchain_list.get(key, None)
             if installer:
+                if verbose:
+                    print(f"+ Found toolchain subclass: {str(installer)}. Invoking.")
                 result = installer.build(base, dirpath, command)
+                if verbose:
+                    print(f"+ Toolchain build result: {result}")
                 return result
             else:
                 print(f"ERROR: unable to find toolchain for specified device")
@@ -170,12 +212,21 @@ class Toolchain:
             print(traceback.format_exc())
             exit(1)
 
-    def build_and_flash(self, base, manufacturer, processor, opsys, version, location, dirpath, command):
+    def build_and_flash(self, base, manufacturer, processor, opsys, version, location, dirpath, command, verbose=False):
         try:
+            if verbose:
+                print(f"+ Looking up build tool for: mfr: {manufacturer} proc: {processor} os: {opsys} ver: {version}")
             key = self._make_key(manufacturer, processor, opsys, version)
+            if verbose:
+                print(f"+ Generated install hash key: {key}")
+
             installer = self.toolchain_list.get(key, None)
             if installer:
+                if verbose:
+                    print(f"+ Found toolchain subclass: {str(installer)}. Invoking.")
                 result = installer.build_and_flash(base, dirpath, command)
+                if verbose:
+                    print(f"+ Toolchain build and flash result: {result}")
                 return result
             else:
                 print(f"ERROR: unable to find toolchain for specified device")
@@ -189,5 +240,5 @@ class Toolchain:
     # This lists the available toolchains this user can install.
     # This returns all the concrete ToolChain classes registered on this system.
     #
-    def list_available(self):
+    def list_available(self, verbose=False):
         return self.toolchain_list
